Initial support for nightly

+++ jenkins-scripts/lib/debbuild-base.bash
@@ -13,6 +13,11 @@ if [ -z $RELEASE_REPO_DIRECTORY ]; then
     RELEASE_REPO_DIRECTORY=ubuntu
 fi;
 
+NIGHTLY_MODE=false
+if [ "${SOURCE_TARBALL_URI}" = "nightly" ]; then
+    NIGHTLY_MODE=true
+fi
+
 ###################################################
 # Boilerplate.
 # DO NOT MODIFY
@@ -111,15 +116,19 @@ pbuilder-dist $DISTRO $ARCH create --othermirror "deb http://packages.ros.org/ro
 # Step 0: Clean up
 rm -rf $WORKSPACE/build
 mkdir -p $WORKSPACE/build
-
-# Step 1: Get the source tarball
 cd $WORKSPACE/build
-wget --quiet -O ${PACKAGE_ALIAS}_$VERSION.orig.tar.bz2 $SOURCE_TARBALL_URI
 
-# Step 3: unpack tarball
-rm -rf $PACKAGE-$VERSION
-tar xf ${PACKAGE_ALIAS}_$VERSION.orig.tar.bz2
-cd $PACKAGE-$VERSION
+# Step 1: Get the source (nightly builds or tarball)
+if ${NIGHTLY_MODE}; then
+  apt-get install -y mercurial
+  hg clone https://bitbucket.org/osrf/$PACKAGE
+  PACKAGE_SRC_BUILD_DIR=$PACKAGE
+else
+  wget --quiet -O ${PACKAGE_ALIAS}_$VERSION.orig.tar.bz2 $SOURCE_TARBALL_URI
+  rm -rf $PACKAGE-$VERSION
+  tar xf ${PACKAGE_ALIAS}_$VERSION.orig.tar.bz2
+  PACKAGE_SRC_BUILD_DIR=$PACKAGE-$VERSION
+fi
 
 # Step 4: add debian/ subdirectory with necessary metadata files to unpacked source tarball
 rm -rf /tmp/$PACKAGE-release
@@ -134,7 +143,7 @@ cd /tmp/$PACKAGE-release/${RELEASE_REPO_DIRECTORY}
 rel_symlinks=\$(find . -maxdepth 1 -type l)
 rel_dirs=\$(find ./* -maxdepth 1 -type d)
 # copy symlinks taking care of existing dirs
-cd $WORKSPACE/build/$PACKAGE-$VERSION
+cd $WORKSPACE/build/\$PACKAGE_SRC_BUILD_DIR
 for sym in \$rel_symlinks; do
    if [ -d \${sym} ]; then
      cp -a /tmp/$PACKAGE-release/${RELEASE_REPO_DIRECTORY}/\${sym}/* \${sym}/
@@ -147,6 +156,18 @@ for dir in \$rel_dirs; do
    cp -a /tmp/$PACKAGE-release/${RELEASE_REPO_DIRECTORY}/\${dir} .
 done
 
+# [nightly] Adjust version in nightly mode
+if $NIGHTLY_MODE; then
+  UPSTREAM_VERSION=1.4.0 # TODO fix this to get latest from changelog
+  TIMESTAMP=\$(date '+%Y%m%d')
+  REV=\$(hg parents --template="{node|short}\n")
+  NIGHTLY_VERSION_SUFFIX=-\${UPSTREAM_VERSION}~hg\${TIMESTAMP}r\${REV}-${RELEASE_VERSION}~${ARCH}
+  NIGHTLY_VERSION=${PACKAGE_ALIAS}-\${NIGHTLY_VERSION_SUFFIX}
+  # Fix the changelog
+  sed -i -e 's:xxxxx:\${NIGHTLY_VERSION}:g' debian/changelog
+  # TODO: Fix CMakeLists.txt ?
+fi
+
 # Step 5: use debuild to create source package
 #TODO: create non-passphrase-protected keys and remove the -uc and -us args to debuild
 debuild -S -uc -us --source-option=--include-binaries
@@ -158,8 +179,14 @@ pbuilder-dist $DISTRO $ARCH build ../*.dsc
 sudo apt-get install -y openssh-client
 cd /var/packages/gazebo/ubuntu
 
-PKG_NAME=${PACKAGE_ALIAS}_${VERSION}-${RELEASE_VERSION}~${DISTRO}_${ARCH}.deb
-DBG_PKG_NAME=${PACKAGE_ALIAS}-dbg_${VERSION}-${RELEASE_VERSION}~${DISTRO}_${ARCH}.deb
+# Set proper package names
+if $NIGHTLY_MODE; then
+  PKG_NAME=\${NIGHTLY_VERSION}.deb
+  DBG_PKG_NAME=${PACKAGE_ALIAS}-dbg\${NIGHTLY_VERSION_SUFFIX}.deb
+else
+  PKG_NAME=${PACKAGE_ALIAS}_${VERSION}-${RELEASE_VERSION}~${DISTRO}_${ARCH}.deb
+  DBG_PKG_NAME=${PACKAGE_ALIAS}-dbg-${VERSION}-${RELEASE_VERSION}~${DISTRO}_${ARCH}.deb
+fi
 
 MAIN_PKGS="/var/lib/jenkins/pbuilder/${DISTRO}-${ARCH}_result/\${PKG_NAME} /var/lib/jenkins/pbuilder/${DISTRO}_result/\${PKG_NAME}"
 DEBUG_PKGS="/var/lib/jenkins/pbuilder/${DISTRO}-${ARCH}_result/\${DBG_PKG_NAME} /var/lib/jenkins/pbuilder/${DISTRO}_result/\${DBG_PKG_NAME}"
